requirements.txt
azure-cosmos
azure-functions
